[Minor] Types for Dropdown component
